Fix x64 compile warnings, except in ext.

+++ lib/include/colorist/raw.h
@@ -13,7 +13,7 @@
 typedef struct clRaw
 {
     uint8_t * ptr;
-    uint32_t size;
+    size_t size;
 } clRaw;
 
 typedef struct clStructArraySchema
@@ -24,11 +24,11 @@ typedef struct clStructArraySchema
 
 struct clContext;
 
-void clRawRealloc(struct clContext * C, clRaw * raw, uint32_t newSize);
+void clRawRealloc(struct clContext * C, clRaw * raw, size_t newSize);
 void clRawClone(struct clContext * C, clRaw * dst, const clRaw * src);
 clBool clRawDeflate(struct clContext * C, clRaw * dst, const clRaw * src);
 char * clRawToBase64(struct clContext * C, clRaw * src);
-void clRawSet(struct clContext * C, clRaw * raw, const uint8_t * data, uint32_t len);
+void clRawSet(struct clContext * C, clRaw * raw, const uint8_t * data, size_t len);
 void clRawFree(struct clContext * C, clRaw * raw);
 clBool clRawReadFile(struct clContext * C, clRaw * raw, const char * filename);
 clBool clRawWriteFile(struct clContext * C, clRaw * raw, const char * filename);
